# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * hollr.chat API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface SuccessEnvelope {
  success: boolean;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus];

export const UserStatus = {
  online: "online",
  idle: "idle",
  dnd: "dnd",
  offline: "offline",
} as const;

export interface User {
  id: string;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  status?: UserStatus;
  /** @nullable */
  customStatus?: string | null;
  createdAt: string;
}

export type UpdateUserRequestStatus =
  (typeof UpdateUserRequestStatus)[keyof typeof UpdateUserRequestStatus];

export const UpdateUserRequestStatus = {
  online: "online",
  idle: "idle",
  dnd: "dnd",
  offline: "offline",
} as const;

export interface UpdateUserRequest {
  /**
   * @minLength 1
   * @maxLength 32
   */
  displayName?: string;
  /** @nullable */
  avatarUrl?: string | null;
  status?: UpdateUserRequestStatus;
  /**
   * @maxLength 128
   * @nullable
   */
  customStatus?: string | null;
}

export interface Server {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  iconUrl?: string | null;
  ownerId: string;
  /** @nullable */
  inviteCode?: string | null;
  memberCount?: number;
  createdAt: string;
}

export interface CreateServerRequest {
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
  /**
   * @maxLength 512
   * @nullable
   */
  description?: string | null;
  /** @nullable */
  iconUrl?: string | null;
}

export interface UpdateServerRequest {
  /**
   * @minLength 1
   * @maxLength 100
   */
  name?: string;
  /**
   * @maxLength 512
   * @nullable
   */
  description?: string | null;
  /** @nullable */
  iconUrl?: string | null;
}

export interface JoinServerRequest {
  /** @minLength 1 */
  inviteCode: string;
}

export interface InviteCode {
  inviteCode: string;
}

export type MemberRole = (typeof MemberRole)[keyof typeof MemberRole];

export const MemberRole = {
  owner: "owner",
  admin: "admin",
  member: "member",
} as const;

export interface Member {
  userId: string;
  serverId: string;
  role: MemberRole;
  joinedAt: string;
  user: User;
}

export type ChannelType = (typeof ChannelType)[keyof typeof ChannelType];

export const ChannelType = {
  text: "text",
  voice: "voice",
} as const;

export interface Channel {
  id: string;
  serverId: string;
  name: string;
  /** @nullable */
  topic?: string | null;
  type: ChannelType;
  position: number;
  createdAt: string;
}

export type CreateChannelRequestType =
  (typeof CreateChannelRequestType)[keyof typeof CreateChannelRequestType];

export const CreateChannelRequestType = {
  text: "text",
  voice: "voice",
} as const;

export interface CreateChannelRequest {
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
  /**
   * @maxLength 1024
   * @nullable
   */
  topic?: string | null;
  type: CreateChannelRequestType;
}

export interface UpdateChannelRequest {
  /**
   * @minLength 1
   * @maxLength 100
   */
  name?: string;
  /**
   * @maxLength 1024
   * @nullable
   */
  topic?: string | null;
}

export interface Attachment {
  id: string;
  objectPath: string;
  name: string;
  contentType: string;
  size: number;
}

export interface Message {
  id: string;
  content: string;
  authorId: string;
  /** @nullable */
  channelId?: string | null;
  /** @nullable */
  dmThreadId?: string | null;
  edited?: boolean;
  attachments?: Attachment[];
  author: User;
  createdAt: string;
  updatedAt?: string;
}

export interface AttachmentInput {
  objectPath: string;
  name: string;
  contentType: string;
  size: number;
}

export interface SendMessageRequest {
  /**
   * @minLength 0
   * @maxLength 4000
   */
  content: string;
  attachments?: AttachmentInput[];
}

export interface EditMessageRequest {
  /**
   * @minLength 1
   * @maxLength 4000
   */
  content: string;
}

export interface DmThread {
  id: string;
  participants: User[];
  lastMessage?: Message | null;
  createdAt: string;
}

export interface OpenDmRequest {
  userId: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type ListMessagesParams = {
  /**
   * Cursor for pagination — fetch messages before this message ID
   */
  before?: string;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};

export type SearchMessagesParams = {
  /**
   * @minLength 1
   */
  q: string;
};

export type ListDmMessagesParams = {
  before?: string;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};