# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: hollr.chat API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /servers/{serverId}/channels:
    get:
      tags: [Channels]
      operationId: listChannels
      summary: List channels in a server
      parameters:
        - $ref: '#/components/parameters/ServerId'
      responses:
        '200':
          description: List of channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Channel'
    post:
      tags: [Channels]
      operationId: createChannel
      summary: Create a channel
      parameters:
        - $ref: '#/components/parameters/ServerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChannelRequest'
      responses:
        '201':
          description: Channel created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /servers/{serverId}/channels/{channelId}:
    patch:
      tags: [Channels]
      operationId: updateChannel
      summary: Update a channel
      parameters:
        - $ref: '#/components/parameters/ServerId'
        - $ref: '#/components/parameters/ChannelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChannelRequest'
      responses:
        '200':
          description: Updated channel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
    delete:
      tags: [Channels]
      operationId: deleteChannel
      summary: Delete a channel
      parameters:
        - $ref: '#/components/parameters/ServerId'
        - $ref: '#/components/parameters/ChannelId'
      responses:
        '200':
          description: Channel deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'

  # Messages
...
  /dms:
    get:
      tags: [DirectMessages]
      operationId: listDmThreads
      summary: List all DM threads for current user
      responses:
        '200':
          description: List of DM threads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DmThread'
    post:
      tags: [DirectMessages]
      operationId: openDmThread
      summary: Open or get a DM thread with another user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenDmRequest'
      responses:
        '200':
          description: DM thread
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DmThread'
  /dms/{threadId}/messages:
    get:
      tags: [DirectMessages]
      operationId: listDmMessages
      summary: List messages in a DM thread
      parameters:
        - $ref: '#/components/parameters/ThreadId'
        - name: before
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: List of DM messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
    post:
      tags: [DirectMessages]
      operationId: sendDmMessage
      summary: Send a message in a DM thread
      parameters:
        - $ref: '#/components/parameters/ThreadId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '201':
          description: DM message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
components:
  parameters:
...
    ServerId:
      name: serverId
      in: path
      required: true
      schema:
        type: string
    ChannelId:
      name: channelId
      in: path
      required: true
      schema:
        type: string
...
    ThreadId:
      name: threadId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    # Auth schemas
...
    User:
      type: object
      required: [id, username, displayName, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: [string, 'null']
        status:
          type: string
          enum: [online, idle, dnd, offline]
          default: offline
        customStatus:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
...
    Channel:
      type: object
      required: [id, serverId, name, type, position, createdAt]
      properties:
        id:
          type: string
        serverId:
          type: string
        name:
          type: string
        topic:
          type: [string, 'null']
        type:
          type: string
          enum: [text, voice]
        position:
          type: integer
        createdAt:
          type: string
          format: date-time
    CreateChannelRequest:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        topic:
          type: [string, 'null']
          maxLength: 1024
        type:
          type: string
          enum: [text, voice]
    UpdateChannelRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        topic:
          type: [string, 'null']
          maxLength: 1024

    # Message schemas
    Message:
      type: object
      required: [id, content, authorId, createdAt, author]
      properties:
        id:
          type: string
        content:
          type: string
        authorId:
          type: string
        channelId:
          type: [string, 'null']
        dmThreadId:
          type: [string, 'null']
        edited:
          type: boolean
          default: false
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
        author:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SendMessageRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 0
          maxLength: 4000
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/AttachmentInput'
...
    Attachment:
      type: object
      required: [id, objectPath, name, contentType, size]
      properties:
        id:
          type: string
        objectPath:
          type: string
        name:
          type: string
        contentType:
          type: string
        size:
          type: integer
    AttachmentInput:
      type: object
      required: [objectPath, name, contentType, size]
      properties:
        objectPath:
          type: string
        name:
          type: string
        contentType:
          type: string
        size:
          type: integer

    # DM schemas
    DmThread:
      type: object
      required: [id, participants, createdAt]
      properties:
        id:
          type: string
        participants:
          type: array
          items:
            $ref: '#/components/schemas/User'
        lastMessage:
          oneOf:
            - $ref: '#/components/schemas/Message'
            - type: 'null'
        createdAt:
          type: string
          format: date-time
    OpenDmRequest:
      type: object
      required: [userId]
      properties:
        userId:
          type: string