# lib/api-zod/src/generated/types/channel.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * hollr.chat API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChannelType } from "./channelType";

export interface Channel {
  id: string;
  serverId: string;
  name: string;
  /** @nullable */
  topic?: string | null;
  type: ChannelType;
  position: number;
  createdAt: Date;
}