# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * hollr.chat API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserStatus } from "./userStatus";

export interface User {
  id: string;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  status?: UserStatus;
  /** @nullable */
  customStatus?: string | null;
  createdAt: Date;
}